requirements.txt
Flask==3.0.0
Werkzeug==3.0.1
mysql-connector-python==8.2.0
PyYAML==6.0